pyproject.toml
[tool.poetry]
name = "nick-raspi-01-bot"
version = "0.0.1"
description = "Telegram Bot for my Raspberry Pi"
readme = "README.md"
authors = ["Nikita Karamov <nick@karamoff.dev>"]
repository = "https://github.com/NickKaramoff/nick-raspi-01-bot"

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^13.0"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[tool.poetry.scripts]
raspi-01-bot = 'nick_raspi_01_bot.main:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
